resource-timing/no-entries-for-cross-origin-css-fetched.sub.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Make sure that resources fetched by cross origin CSS are not in the timeline.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<link rel=stylesheet id=cross_origin_style href="//{{domains[www1]}}:{{ports[http][1]}}{{location[path]}}/../resources/nested.css">
<script>
<<<<<<< HEAD
<<<<<<< HEAD
  let t = async_test("Make sure that resources fetched by cross origin CSS are not in the timeline.");
  window.onload = function() {
    // A timeout is needed as PerformanceObserver is not guaranteed to run before onload triggered.
    t.step_timeout(function() {
          let url = (new URL(document.getElementById("cross_origin_style").href));
          let prefix = url.protocol + "//" + url.host;
          assert_equals(performance.getEntriesByName(prefix + "/resource-timing/resources/resource_timing_test0.css?id=n1").length, 0, "Import is in timeline");
          assert_equals(performance.getEntriesByName(prefix + "/fonts/Ahem.ttf?id=n1").length, 0, "Font is in timeline");
          assert_equals(performance.getEntriesByName(prefix + "/resource-timing/resources/blue.png?id=n1").length, 0, "Image is in timeline");
<<<<<<< HEAD
=======
=======
  let t = async_test("Make sure that resources fetched by cross origin CSS are not in the timeline.");
>>>>>>> 36672595d0... Resolved lint complaints
  window.onload = function() {
    // A timeout is needed as PerformanceObserver is not guaranteed to run before onload triggered.
    t.step_timeout(function() {
          let url = (new URL(document.getElementById("cross_origin_style").href));
          let prefix = url.protocol + "//" + url.host;
          assert_equals(performance.getEntriesByName(prefix + "/fonts/Ahem.ttf?id=n1").length, 0, "Font is in timeline");
          assert_equals(performance.getEntriesByName(prefix + "/resource-timing/resources/blue.png?id=n1").length, 0, "Image is in timeline");
<<<<<<< HEAD
      }, "Make sure that CSS subresources are not included");
>>>>>>> 3c0ba54dd2... Test that resources fetched by cross-origin CSS do not appear in the timeline
=======
>>>>>>> 36672595d0... Resolved lint complaints
=======
          t.done();
>>>>>>> f1012e750d... Rebase and add import
    },0)
  };
</script>
<ol>Some content</ol>
</body>
