third_party/blink/web_tests/platform/generic/external/wpt/resource-timing/no-entries-for-cross-origin-css-fetched.sub-expected.txt

This is a testharness.js-based test.
FAIL Make sure that resources fetched by cross origin CSS are not in the timeline. assert_equals: Import should not be in timeline expected 0 but got 1
Harness: the test ran to completion.
